requirements.txt
fastapi>=0.104.0,<1.0.0
uvicorn>=0.24.0,<1.0.0
paramiko>=3.4.0,<4.0.0
pyyaml>=6.0,<7.0
pydantic>=2.5.0,<3.0.0
jinja2>=3.1.0,<4.0.0
markdown>=3.5.0,<4.0.0
